backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
boto3==1.34.40
mangum==0.17.0
python-dotenv==1.0.1
circuitbreaker
pytest
numpy
scikit-learn
PyJWT
python-multipart
python-json-logger
google-genai
